requirements.txt
tastyworks-api
pymongo